refactor(core): split tool decision truth layer
